check role in some service get current user in service that need add new service for register user

+++ src/main/java/com/snapp/dangidongi/controller/UserController.java
@@ -3,6 +3,7 @@ package com.snapp.dangidongi.controller;
 import com.snapp.dangidongi.common.Url;
 import com.snapp.dangidongi.mapper.UserMapper;
 import com.snapp.dangidongi.model.UserModel;
+import com.snapp.dangidongi.security.UserDetail;
 import com.snapp.dangidongi.service.UserService;
 import lombok.AllArgsConstructor;
 import lombok.SneakyThrows;
@@ -11,6 +12,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -19,12 +23,12 @@ import java.net.URI;
 
 @RestController
 @AllArgsConstructor
-//TODO: limit access to this service in authorization
 public class UserController {
 
     private final UserService userService;
     private final UserMapper userMapper;
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @PostMapping(value = Url.USERS, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Long> createUser(@RequestBody @Validated UserModel user) {
         var id = userService.save(user).getId();
@@ -36,10 +40,23 @@ public class UserController {
         return ResponseEntity.created(uri).build();
     }
 
+    @PostMapping(value = Url.REGISTER, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Long> registerUser(@RequestBody @Validated UserModel user) {
+        var id = userService.save(user).getId();
+        URI uri = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(id)
+                .toUri();
+        return ResponseEntity.created(uri).build();
+    }
 
     @SneakyThrows
     @GetMapping(value = Url.USERS_ID, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UserModel> getUserById(@PathVariable Long id) {
+    public ResponseEntity<UserModel> getUserById(@AuthenticationPrincipal UserDetail userDetails, @PathVariable Long id) {
+        if (!userDetails.getUser().getId().equals(id)) {
+            throw new AccessDeniedException("can not get user by id " + id);
+        }
         UserModel model = userMapper.userEntityToUserModel(userService.findById(id));
         return ResponseEntity.ok(model);
     }
@@ -59,7 +76,10 @@ public class UserController {
 
 
     @DeleteMapping(value = Url.USERS_ID, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UserModel> deleteUserById(@PathVariable Long id) {
+    public ResponseEntity<UserModel> deleteUserById(@AuthenticationPrincipal UserDetail userDetails, @PathVariable Long id) {
+        if (!userDetails.getUser().getId().equals(id)) {
+            throw new AccessDeniedException("can not user by id " + id);
+        }
         userService.deleteById(id);
         return ResponseEntity.ok().build();
     }
